input str->int fix

+++ test.py
@@ -34,7 +34,7 @@ def servo(x,value):
     if value>100 or value<0:
         print("Wrong value! (expect 0-100)")
     else:
-        pwm.set_pwm(x,0,((servo_min[x])+((  (servo_max[x]-servo_min[x])/100) * value ) )  )
+        pwm.set_pwm(x,0,int((servo_min[x])+((  (servo_max[x]-servo_min[x])/100) * value ) )  )
 
 
 if __name__ == "__main__":
